Implement a more sane memmap_chomp function
Just return the size without the trailing whitespace characters
instead of changing the stored size information from the actual
mmap(2) call. This results in the correct len passed to unmap(2)
when freeing the mapped file.

+++ helper.h
@@ -30,7 +30,7 @@ void		 decode_string(char *);
 struct memmap	*memmap_new(const char *);
 struct memmap	*memmap_new_at(int, const char *);
 void		 memmap_free(struct memmap *);
-int		 memmap_chomp(struct memmap *);
+size_t		 memmap_chomp(struct memmap *);
 
 
 #endif // __HELPER_H__
